MAINT: Code reorganization
