+++ dtool_lookup_gui/Dependencies.py
@@ -25,7 +25,7 @@ import json
 import logging
 
 from . import is_uuid
-from .SimpleGraph import SimpleGraph
+from .widgets.simple_graph import SimpleGraph
 
 
 logger = logging.getLogger(__name__)
