+++ dtool_lookup_gui/views/settings_dialog.ui
@@ -14,14 +14,16 @@
       <object class="GtkBox">
         <property name="visible">True</property>
         <property name="can-focus">False</property>
+        <property name="margin-start">12</property>
+        <property name="margin-end">12</property>
+        <property name="margin-top">12</property>
+        <property name="margin-bottom">12</property>
         <property name="orientation">vertical</property>
         <child>
           <object class="GtkLabel">
             <property name="visible">True</property>
             <property name="can-focus">False</property>
             <property name="halign">start</property>
-            <property name="margin-start">64</property>
-            <property name="margin-top">32</property>
             <property name="label" translatable="yes">Lookup configuration</property>
             <attributes>
               <attribute name="weight" value="bold"/>
@@ -37,8 +39,6 @@
           <object class="GtkFrame">
             <property name="visible">True</property>
             <property name="can-focus">False</property>
-            <property name="margin-start">64</property>
-            <property name="margin-end">64</property>
             <property name="margin-top">12</property>
             <property name="label-xalign">0</property>
             <property name="shadow-type">in</property>
@@ -216,8 +216,6 @@
             <property name="visible">True</property>
             <property name="can-focus">False</property>
             <property name="halign">start</property>
-            <property name="margin-start">64</property>
-            <property name="margin-end">64</property>
             <property name="margin-top">32</property>
             <property name="label" translatable="yes">Endpoint configuration</property>
             <attributes>
@@ -234,10 +232,7 @@
           <object class="GtkFrame">
             <property name="visible">True</property>
             <property name="can-focus">False</property>
-            <property name="margin-start">64</property>
-            <property name="margin-end">64</property>
             <property name="margin-top">12</property>
-            <property name="margin-bottom">32</property>
             <property name="label-xalign">0</property>
             <property name="shadow-type">in</property>
             <child>
@@ -245,6 +240,9 @@
                 <property name="visible">True</property>
                 <property name="can-focus">False</property>
                 <property name="selection-mode">none</property>
+                <child type="placeholder">
+                  <placeholder/>
+                </child>
               </object>
             </child>
             <child type="label_item">
